Refactor artifact upload for improved logging and retention
