+++ .github/workflows/Agents.yml
@@ -39,9 +39,15 @@ jobs:
         env:
           JOB_STATUS: ${{ job.status }}
         run: npx tsx helpers/slack.ts
-      - name: Upload reports artifacts
-        if: always()
+      - name: Upload test artifacts
         uses: actions/upload-artifact@v4
         with:
-          name: logs_${{ matrix.environment }}
-          path: logs/
+          name: logs-${{ matrix.environment }}
+          path: |
+            logs/
+            .data/
+            .env
+          if-no-files-found: warn
+          overwrite: true
+          include-hidden-files: true
+          retention-days: 90
