+++ .github/workflows/Delivery.yml
@@ -44,9 +44,15 @@ jobs:
         run: npx tsx helpers/slack.ts
       - name: Ensure logs directory exists
         run: mkdir -p logs
-      - name: Upload reports artifacts
-        if: always()
+      - name: Upload test artifacts
         uses: actions/upload-artifact@v4
         with:
-          name: logs_${{ matrix.environment }}
-          path: logs/
+          name: logs-${{ matrix.environment }}
+          path: |
+            logs/
+            .data/
+            .env
+          if-no-files-found: warn
+          overwrite: true
+          include-hidden-files: true
+          retention-days: 90
