+++ .github/workflows/StressGroup.yml
@@ -63,7 +63,7 @@ jobs:
       - name: Upload test artifacts
         uses: actions/upload-artifact@v4
         with:
-          name: logs
+          name: logs-${{ matrix.environment }}
           path: |
             logs/
             .data/
